Moved startup settings to preferences
