+++ Files/Dialogs/SettingsDialog.xaml.cs
@@ -47,11 +47,10 @@ namespace Files.Dialogs
             _ = selectedItemTag switch
             {
                 0 => SettingsContentFrame.Navigate(typeof(Appearance)),
-                1 => SettingsContentFrame.Navigate(typeof(OnStartup)),
-                2 => SettingsContentFrame.Navigate(typeof(Preferences)),
-                3 => SettingsContentFrame.Navigate(typeof(Multitasking)),
-                4 => SettingsContentFrame.Navigate(typeof(Experimental)),
-                5 => SettingsContentFrame.Navigate(typeof(About)),
+                1 => SettingsContentFrame.Navigate(typeof(Preferences)),
+                2 => SettingsContentFrame.Navigate(typeof(Multitasking)),
+                3 => SettingsContentFrame.Navigate(typeof(Experimental)),
+                4 => SettingsContentFrame.Navigate(typeof(About)),
                 _ => SettingsContentFrame.Navigate(typeof(Appearance))
             };
         }
