+++ Files/Services/Implementation/PreferencesSettingsService.cs
@@ -1,4 +1,5 @@
 ﻿using Files.Models.JsonSettings;
+using System.Collections.Generic;
 
 namespace Files.Services.Implementation
 {
@@ -26,6 +27,10 @@ namespace Files.Services.Implementation
                 case nameof(AreLayoutPreferencesPerFolder):
                 case nameof(AdaptiveLayoutEnabled):
                 case nameof(AreFileTagsEnabled):
+                case nameof(OpenSpecificPageOnStartup):
+                case nameof(ContinueLastSessionOnStartUp):
+                case nameof(OpenNewTabOnStartup):
+                case nameof(AlwaysOpenNewInstance):
                     Microsoft.AppCenter.Analytics.Analytics.TrackEvent($"{nameof(e.settingName)} {e.newValue}");
                     break;
             }
@@ -102,6 +107,48 @@ namespace Files.Services.Implementation
         public bool AreFileTagsEnabled
         {
             get => Get(false);
+            set => Set(value);
+        }
+
+        public bool OpenSpecificPageOnStartup
+        {
+            get => Get(false);
+            set => Set(value);
+        }
+
+        public string OpenSpecificPageOnStartupPath
+        {
+            get => Get(string.Empty);
+            set => Set(value);
+        }
+
+        public bool ContinueLastSessionOnStartUp
+        {
+            get => Get(false);
+            set => Set(value);
+        }
+
+        public bool OpenNewTabOnStartup
+        {
+            get => Get(true);
+            set => Set(value);
+        }
+
+        public bool AlwaysOpenNewInstance
+        {
+            get => Get(false);
+            set => Set(value);
+        }
+
+        public List<string> TabsOnStartupList
+        {
+            get => Get<List<string>>(null);
+            set => Set(value);
+        }
+
+        public List<string> LastSessionTabList
+        {
+            get => Get<List<string>>(null);
             set => Set(value);
         }
     }
