+++ Files/ViewModels/SettingsViewModels/PreferencesViewModel.cs
@@ -1,6 +1,8 @@
-﻿using Files.Controllers;
+﻿using Files.Common;
+using Files.Controllers;
 using Files.DataModels;
 using Files.Enums;
+using Files.Filesystem;
 using Files.Helpers;
 using Files.Services;
 using Microsoft.Toolkit.Mvvm.ComponentModel;
@@ -10,11 +12,16 @@ using Microsoft.Toolkit.Uwp;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Collections.Specialized;
+using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Input;
 using Windows.Foundation.Collections;
 using Windows.Storage;
+using Windows.Storage.AccessCache;
+using Windows.Storage.Pickers;
 using Windows.System;
+using static Files.Helpers.MenuFlyoutHelper;
 
 namespace Files.ViewModels.SettingsViewModels
 {
@@ -28,6 +35,9 @@ namespace Files.ViewModels.SettingsViewModels
         private int selectedDateFormatIndex = (int)Enum.Parse(typeof(TimeStyle), App.AppSettings.DisplayedTimeStyle.ToString());
         private List<Terminal> terminals;
         private bool disposed;
+        private int selectedPageIndex = -1;
+        private bool isPageListEditEnabled;
+        private ReadOnlyCollection<IMenuFlyoutItem> addFlyoutItemsSource;
 
         public ICommand EditTerminalApplicationsCommand { get; }
 
@@ -43,6 +53,251 @@ namespace Files.ViewModels.SettingsViewModels
 
             EditTerminalApplicationsCommand = new AsyncRelayCommand(LaunchTerminalsConfigFile);
             App.TerminalController.ModelChanged += ReloadTerminals;
+
+            if (UserSettingsService.PreferencesSettingsService.TabsOnStartupList != null)
+            {
+                PagesOnStartupList = new ObservableCollection<PageOnStartupViewModel>(UserSettingsService.PreferencesSettingsService.TabsOnStartupList.Select((p) => new PageOnStartupViewModel(p)));
+            }
+            else
+            {
+                PagesOnStartupList = new ObservableCollection<PageOnStartupViewModel>();
+            }
+
+            PagesOnStartupList.CollectionChanged += PagesOnStartupList_CollectionChanged;
+
+            var recentsItem = new MenuFlyoutSubItemViewModel("JumpListRecentGroupHeader".GetLocalized());
+            recentsItem.Items.Add(new MenuFlyoutItemViewModel("SidebarHome".GetLocalized(), "Home".GetLocalized(), AddPageCommand));
+            PopulateRecentItems(recentsItem).ContinueWith(_ =>
+            {
+                AddFlyoutItemsSource = new ReadOnlyCollection<IMenuFlyoutItem>(new IMenuFlyoutItem[] {
+                    new MenuFlyoutItemViewModel("Browse".GetLocalized(), null, AddPageCommand),
+                    recentsItem,
+                });
+            }, TaskScheduler.FromCurrentSynchronizationContext());
+        }
+
+
+        private async Task PopulateRecentItems(MenuFlyoutSubItemViewModel menu)
+        {
+            bool hasRecents = false;
+            menu.Items.Add(new MenuFlyoutSeparatorViewModel());
+
+            try
+            {
+                var mostRecentlyUsed = StorageApplicationPermissions.MostRecentlyUsedList;
+
+                foreach (AccessListEntry entry in mostRecentlyUsed.Entries)
+                {
+                    string mruToken = entry.Token;
+                    var added = await FilesystemTasks.Wrap(async () =>
+                    {
+                        IStorageItem item = await mostRecentlyUsed.GetItemAsync(mruToken, AccessCacheOptions.FastLocationsOnly);
+                        if (item.IsOfType(StorageItemTypes.Folder))
+                        {
+                            menu.Items.Add(new MenuFlyoutItemViewModel(item.Name, string.IsNullOrEmpty(item.Path) ? entry.Metadata : item.Path, AddPageCommand));
+                            hasRecents = true;
+                        }
+                    });
+                    if (added == FileSystemStatusCode.Unauthorized)
+                    {
+                        // Skip item until consent is provided
+                    }
+                    // Exceptions include but are not limited to:
+                    // COMException, FileNotFoundException, ArgumentException, DirectoryNotFoundException
+                    // 0x8007016A -> The cloud file provider is not running
+                    // 0x8000000A -> The data necessary to complete this operation is not yet available
+                    // 0x80004005 -> Unspecified error
+                    // 0x80270301 -> ?
+                    else if (!added)
+                    {
+                        await FilesystemTasks.Wrap(() =>
+                        {
+                            mostRecentlyUsed.Remove(mruToken);
+                            return Task.CompletedTask;
+                        });
+                        System.Diagnostics.Debug.WriteLine(added.ErrorCode);
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                App.Logger.Info(ex, "Could not fetch recent items");
+            }
+
+            if (!hasRecents)
+            {
+                menu.Items.RemoveAt(menu.Items.Count - 1);
+            }
+        }
+
+        private void PagesOnStartupList_CollectionChanged(object sender, NotifyCollectionChangedEventArgs e)
+        {
+            if (PagesOnStartupList.Count() > 0)
+            {
+                UserSettingsService.PreferencesSettingsService.TabsOnStartupList = PagesOnStartupList.Select((p) => p.Path).ToList();
+            }
+            else
+            {
+                UserSettingsService.PreferencesSettingsService.TabsOnStartupList = null;
+            }
+        }
+
+        public bool OpenNewTabPageOnStartup
+        {
+            get => UserSettingsService.PreferencesSettingsService.OpenNewTabOnStartup;
+            set
+            {
+                if (value != UserSettingsService.PreferencesSettingsService.OpenNewTabOnStartup)
+                {
+                    UserSettingsService.PreferencesSettingsService.OpenNewTabOnStartup = value;
+                    OnPropertyChanged();
+                }
+            }
+        }
+
+        public bool ContinueLastSessionOnStartUp
+        {
+            get => UserSettingsService.PreferencesSettingsService.ContinueLastSessionOnStartUp;
+            set
+            {
+                if (value != UserSettingsService.PreferencesSettingsService.ContinueLastSessionOnStartUp)
+                {
+                    UserSettingsService.PreferencesSettingsService.ContinueLastSessionOnStartUp = value;
+                    OnPropertyChanged();
+                }
+            }
+        }
+
+        public bool OpenASpecificPageOnStartup
+        {
+            get => UserSettingsService.PreferencesSettingsService.OpenSpecificPageOnStartup;
+            set
+            {
+                if (value != UserSettingsService.PreferencesSettingsService.OpenSpecificPageOnStartup)
+                {
+                    UserSettingsService.PreferencesSettingsService.OpenSpecificPageOnStartup = value;
+                    OnPropertyChanged();
+                }
+            }
+        }
+
+        public ObservableCollection<PageOnStartupViewModel> PagesOnStartupList { get; set; }
+
+        public int SelectedPageIndex
+        {
+            get => selectedPageIndex;
+            set
+            {
+                if (SetProperty(ref selectedPageIndex, value))
+                {
+                    IsPageListEditEnabled = value >= 0;
+                }
+            }
+        }
+
+        public bool IsPageListEditEnabled
+        {
+            get => isPageListEditEnabled;
+            set => SetProperty(ref isPageListEditEnabled, value);
+        }
+
+        public ReadOnlyCollection<IMenuFlyoutItem> AddFlyoutItemsSource
+        {
+            get => addFlyoutItemsSource;
+            set => SetProperty(ref addFlyoutItemsSource, value);
+        }
+
+        public RelayCommand ChangePageCommand => new RelayCommand(ChangePage);
+        public RelayCommand RemovePageCommand => new RelayCommand(RemovePage);
+        public RelayCommand<string> AddPageCommand => new RelayCommand<string>(AddPage);
+
+        public bool AlwaysOpenANewInstance
+        {
+            get => UserSettingsService.PreferencesSettingsService.AlwaysOpenNewInstance;
+            set
+            {
+                if (value != UserSettingsService.PreferencesSettingsService.AlwaysOpenNewInstance)
+                {
+                    UserSettingsService.PreferencesSettingsService.AlwaysOpenNewInstance = value;
+                    ApplicationData.Current.LocalSettings.Values["AlwaysOpenANewInstance"] = value; // Needed in Program.cs
+                    OnPropertyChanged();
+                }
+            }
+        }
+
+        private async void ChangePage()
+        {
+            var folderPicker = new FolderPicker();
+            folderPicker.FileTypeFilter.Add("*");
+            StorageFolder folder = await folderPicker.PickSingleFolderAsync();
+
+            if (folder != null)
+            {
+                if (SelectedPageIndex >= 0)
+                {
+                    PagesOnStartupList[SelectedPageIndex] = new PageOnStartupViewModel(folder.Path);
+                }
+            }
+        }
+
+        private void RemovePage()
+        {
+            int index = SelectedPageIndex;
+            if (index >= 0)
+            {
+                PagesOnStartupList.RemoveAt(index);
+                if (index > 0)
+                {
+                    SelectedPageIndex = index - 1;
+                }
+                else if (PagesOnStartupList.Count > 0)
+                {
+                    SelectedPageIndex = 0;
+                }
+            }
+        }
+
+        private async void AddPage(string path = null)
+        {
+            if (string.IsNullOrWhiteSpace(path))
+            {
+                var folderPicker = new FolderPicker();
+                folderPicker.FileTypeFilter.Add("*");
+
+                var folder = await folderPicker.PickSingleFolderAsync();
+                if (folder != null)
+                {
+                    path = folder.Path;
+                }
+            }
+
+            if (path != null && PagesOnStartupList != null)
+            {
+                PagesOnStartupList.Add(new PageOnStartupViewModel(path));
+            }
+        }
+
+        public class PageOnStartupViewModel
+        {
+            public string Text
+            {
+                get
+                {
+                    if (Path == "Home".GetLocalized())
+                    {
+                        return "SidebarHome".GetLocalized();
+                    }
+                    if (Path == CommonPaths.RecycleBinPath)
+                    {
+                        return ApplicationData.Current.LocalSettings.Values.Get("RecycleBin_Title", "Recycle Bin");
+                    }
+                    return Path;
+                }
+            }
+
+            public string Path { get; }
+
+            internal PageOnStartupViewModel(string path) => Path = path;
         }
 
         private void ReloadTerminals(TerminalController controller)
