--- Files/Views/SettingsPages/OnStartup.xaml
@@ -1,148 +1,0 @@
-<Page
-    x:Class="Files.SettingsPages.OnStartup"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:converters="using:Microsoft.Toolkit.Uwp.UI.Converters"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:helpers="using:Files.Helpers"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:settingsviewmodels="using:Files.ViewModels.SettingsViewModels"
-    mc:Ignorable="d">
-    <Page.DataContext>
-        <settingsviewmodels:OnStartupViewModel />
-    </Page.DataContext>
-
-    <Page.Resources>
-        <converters:BoolNegationConverter x:Key="BoolNegationConverter" />
-    </Page.Resources>
-
-    <Grid>
-        <StackPanel Orientation="Vertical" Spacing="20">
-            <StackPanel.ChildrenTransitions>
-                <TransitionCollection>
-                    <EntranceThemeTransition />
-                </TransitionCollection>
-            </StackPanel.ChildrenTransitions>
-
-            <TextBlock
-                x:Uid="SettingsOnStartupTitle"
-                FontSize="20"
-                FontWeight="Medium"
-                Text="On Startup" />
-
-            <TextBlock
-                x:Uid="SettingsWhenLaunchingFiles"
-                FontSize="16"
-                Text="When launching Files" />
-
-            <RadioButton
-                x:Uid="SettingsOnStartupOpenANewTab"
-                Grid.Column="0"
-                VerticalAlignment="Center"
-                Content="Open a new tab"
-                GroupName="StartupPageSettingRadioGroup"
-                IsChecked="{Binding OpenNewTabPageOnStartup, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
-
-            <RadioButton
-                x:Uid="SettingsOnStartupContinueWhereYouLeftOff"
-                Grid.Column="0"
-                VerticalAlignment="Center"
-                Content="Continue where you left off"
-                GroupName="StartupPageSettingRadioGroup"
-                IsChecked="{Binding ContinueLastSessionOnStartUp, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
-
-            <RadioButton
-                x:Name="OpenSpecificPage"
-                x:Uid="SettingsOnStartupOpenASpecificPage"
-                Grid.Column="0"
-                VerticalAlignment="Center"
-                Content="Open a specific page or pages"
-                GroupName="StartupPageSettingRadioGroup"
-                IsChecked="{Binding OpenASpecificPageOnStartup, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
-
-            <Grid HorizontalAlignment="Stretch" Visibility="{x:Bind OpenSpecificPage.IsChecked, Mode=OneWay}">
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="*" />
-                    <ColumnDefinition Width="Auto" />
-                </Grid.ColumnDefinitions>
-                <Grid.RowDefinitions>
-                    <RowDefinition />
-                    <RowDefinition />
-                </Grid.RowDefinitions>
-                <TextBlock
-                    x:Uid="SettingsOnStartupPages"
-                    HorizontalAlignment="Left"
-                    VerticalAlignment="Center"
-                    FontSize="16"
-                    Text="Pages: " />
-                <CommandBar
-                    Grid.Column="1"
-                    Background="Transparent"
-                    DefaultLabelPosition="Right"
-                    IsOpen="False">
-                    <AppBarButton
-                        x:Uid="AppBarButtonAdd"
-                        Icon="Add"
-                        Label="Add">
-                        <AppBarButton.Flyout>
-                            <MenuFlyout helpers:MenuFlyoutHelper.ItemsSource="{Binding AddFlyoutItemsSource}" Placement="Bottom" />
-                        </AppBarButton.Flyout>
-                    </AppBarButton>
-                    <AppBarButton
-                        x:Uid="AppBarButtonEdit"
-                        Command="{Binding ChangePageCommand}"
-                        Icon="Edit"
-                        IsEnabled="{Binding IsPageListEditEnabled}"
-                        Label="Edit" />
-                    <AppBarButton
-                        x:Uid="AppBarButtonRemove"
-                        Command="{Binding RemovePageCommand}"
-                        Icon="Delete"
-                        IsEnabled="{Binding IsPageListEditEnabled}"
-                        Label="Remove" />
-                </CommandBar>
-
-                <ListView
-                    x:Name="PagesList"
-                    Grid.Row="1"
-                    Grid.ColumnSpan="2"
-                    Margin="0,10"
-                    HorizontalAlignment="Stretch"
-                    AllowDrop="True"
-                    CanReorderItems="True"
-                    IsItemClickEnabled="True"
-                    ItemsSource="{Binding PagesOnStartupList, Mode=TwoWay}"
-                    SelectedIndex="{Binding SelectedPageIndex, Mode=TwoWay}"
-                    SelectionMode="Single">
-                    <ListView.ItemTemplate>
-                        <DataTemplate>
-                            <TextBlock
-                                Text="{Binding Text}"
-                                TextTrimming="CharacterEllipsis"
-                                ToolTipService.ToolTip="{Binding Text}" />
-                        </DataTemplate>
-                    </ListView.ItemTemplate>
-                </ListView>
-            </Grid>
-
-            <TextBlock
-                x:Uid="SettingsOnStartupNewInstanceBehavior"
-                FontSize="16"
-                Text="When opening a folder in Files" />
-
-            <RadioButton
-                x:Uid="SettingsOnStartupLaunchNewTab"
-                VerticalAlignment="Center"
-                Content="Open folder in a new tab"
-                GroupName="StartupPageLaunchRadioGroup"
-                IsChecked="{Binding AlwaysOpenANewInstance, Mode=TwoWay, Converter={StaticResource BoolNegationConverter}}" />
-
-            <RadioButton
-                x:Uid="SettingsOnStartupLaunchNewInstance"
-                VerticalAlignment="Center"
-                Content="Always open a new instance"
-                GroupName="StartupPageLaunchRadioGroup"
-                IsChecked="{Binding AlwaysOpenANewInstance, Mode=TwoWay}" />
-        </StackPanel>
-    </Grid>
-</Page>
