--- Files/Views/SettingsPages/OnStartup.xaml.cs
@@ -1,12 +1,0 @@
-﻿using Windows.UI.Xaml.Controls;
-
-namespace Files.SettingsPages
-{
-    public sealed partial class OnStartup : Page
-    {
-        public OnStartup()
-        {
-            InitializeComponent();
-        }
-    }
-}
